Implementado regras de autorização no Project

+++ app/Http/Controllers/ProjectController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 
 use CodeProject\Repositories\ProjectRepository;
 use CodeProject\Services\ProjectService;
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
 
 class ProjectController extends Controller
 {
@@ -64,6 +65,10 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
+        if ($this->checkProjectOwner($id) == false) {
+            return ['error' => 'Access Forbidden'];
+        }
+
         return $this->service->find($id);
     }
 
@@ -87,6 +92,10 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if ($this->checkProjectOwner($id) == false) {
+            return ['error' => 'Access Forbidden'];
+        }
+
         return $this->service->update($request->all(), $id);
     }
 
@@ -98,6 +107,10 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
+        if ($this->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access Forbidden'];
+        }
+
         return $this->service->delete($id);
     }
 
@@ -119,5 +132,28 @@ class ProjectController extends Controller
     public function isMember(Request $request, $id, $userId)
     {
         return $this->service->isMember($id, $userId);
+    }
+
+    private function checkProjectOwner($projectId)
+    {
+        $userId = Authorizer::getResourceOwnerId();
+
+        return $this->repository->isOwner($projectId, $userId);
+    }
+
+    private function checkProjectMember($projectId)
+    {
+        $userId = Authorizer::getResourceOwnerId();
+
+        return $this->repository->hasMember($projectId, $userId);
+    }
+
+    private function checkProjectPermissions($projectId)
+    {
+        if ($this->checkProjectOwner($projectId) or $this->checkProjectMember($projectId)) {
+            return true;
+        }
+
+        return false;
     }
 }
